fix(build): stop shipping 12MB of unreachable Monaco workers in the VSIX
The 0.40.1 package jumped from 8.6MB to 13.4MB. All of it was Monaco
language-worker bundles that no shipped code loads.

The extension host has no Monaco: it diffs through VS Code's own
`vscode.diff`. Monaco belongs to the desktop renderer, which loads it for
the editor pane and the inline diff view. But Vite emits a worker bundle
for every `import('...?worker')` expression it finds in the module graph,
at transform time, *before* tree-shaking. So when ToolEditRequestPanel
started importing `monacoLanguageForPath` — a pure path -> language-id
switch that happened to live in `monacoLoader.ts` — the five
`monaco-editor/**?worker` imports inside `loadMonaco` became "visible" and
Vite wrote all of them out, even though rolldown then dropped `loadMonaco`
itself as unreachable and Monaco's core never entered the bundle at all.

Cost: 11.8MB in `dist/assets/` (ts.worker alone is 8.5MB) plus 9.4MB
beside the single-file trace viewer, which shares the progressView graph
and cannot inline `?worker` chunks. Neither artifact references any of
those files at runtime.

Move the table to `@shared/monaco/monacoLanguage`, a leaf module with no
`monaco-editor` import, and point both callers at it. It stays the single
owner of the mapping. The desktop is unaffected — it still imports
`loadMonaco` and ships Monaco as before.

Verified: `resources/traceViewer/` is one 5.12MB index.html again (was 6
files, 14.14MB), `dist/assets/` is gone, `dist/progressView/bundle.js` is
byte-for-byte the same size, VSIX back to 8.27MB / 172 files.

Also pin it in `check:vsix-contents`, which already runs in CI and
release: a packaged Monaco worker is always a bundling accident, since
nothing in the extension can load one. Verified the check fails on the
exact entry names the regressed build produced.
