+++ src/shared/monaco/monacoLoader.ts
@@ -6,6 +6,10 @@
 // copies would race to install it and whichever lost would silently hand the
 // wrong worker to a language service. One loader, one `MonacoEnvironment`
 // assignment, shared promise cache.
+//
+// Everything in here reaches `monaco-editor`. The path -> language-id table
+// lives in monacoLanguage.ts precisely so callers that only need a language id
+// do not import this module — see the comment there for what that costs.
 
 import { DESKTOP_THEME_KIND } from '@shared/schemas';
 
@@ -219,95 +223,4 @@ export function monacoThemeForHostTheme(
   if (themeKind === DESKTOP_THEME_KIND.LIGHT) return 'vs';
   if (themeKind === DESKTOP_THEME_KIND.HIGH_CONTRAST) return 'hc-black';
   return 'vs-dark';
-}
-
-/**
- * Monaco language id for a file path, covering the extensions TeXRA users
- * actually open. Monaco's own registry is keyed by language id, not extension,
- * so this mapping has to live on our side. This is the single owner of that
- * mapping across every host — do not add a second one.
- */
-export function monacoLanguageForPath(filePath: string): string {
-  const name = filePath.replaceAll('\\', '/').split('/').at(-1) ?? '';
-  const lowerName = name.toLowerCase();
-  if (lowerName === 'dockerfile') return 'dockerfile';
-  if (lowerName === 'makefile') return 'makefile';
-  const dotIndex = name.lastIndexOf('.');
-  const extension =
-    dotIndex === -1 ? '' : name.slice(dotIndex + 1).toLowerCase();
-  switch (extension) {
-    case 'tex':
-    case 'sty':
-    case 'cls':
-    case 'bbl':
-      return 'latex';
-    case 'bib':
-      return 'bibtex';
-    case 'md':
-    case 'markdown':
-      return 'markdown';
-    case 'mdx':
-      return 'mdx';
-    case 'ts':
-    case 'mts':
-    case 'cts':
-    case 'tsx':
-      return 'typescript';
-    case 'js':
-    case 'mjs':
-    case 'cjs':
-    case 'jsx':
-      return 'javascript';
-    case 'json':
-    case 'jsonc':
-      return 'json';
-    case 'py':
-      return 'python';
-    case 'lean':
-      return 'lean';
-    case 'yaml':
-    case 'yml':
-      return 'yaml';
-    case 'sh':
-    case 'bash':
-    case 'zsh':
-      return 'shell';
-    case 'css':
-      return 'css';
-    case 'scss':
-      return 'scss';
-    case 'less':
-      return 'less';
-    case 'html':
-    case 'htm':
-      return 'html';
-    case 'toml':
-      return 'ini';
-    case 'xml':
-      return 'xml';
-    case 'sql':
-      return 'sql';
-    case 'rs':
-      return 'rust';
-    case 'go':
-      return 'go';
-    case 'java':
-      return 'java';
-    case 'cs':
-      return 'csharp';
-    case 'php':
-      return 'php';
-    case 'rb':
-      return 'ruby';
-    case 'c':
-    case 'h':
-      return 'c';
-    case 'cpp':
-    case 'cc':
-    case 'cxx':
-    case 'hpp':
-      return 'cpp';
-    default:
-      return 'plaintext';
-  }
 }
